Открывается и закрывается ч2
